apps/agents/planning-ingestion/pyproject.toml
[project]
name = "planning-ingestion"
version = "0.1.0"
description = "Google ADK agent for crawling Milwaukee planning documents"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-adk>=0.5.0",
    "google-genai>=1.0.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "opik>=0.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-httpx>=0.30.0",
]

[project.scripts]
planning-sync = "main:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
